--- client/talk-plugin-flags/index.js
@@ -1,1 +1,0 @@
-export { default as FlagComment } from './components/FlagComment';
